+++ pytket/extensions/qiskit/backends/ibm.py
@@ -32,9 +32,9 @@ from typing import (
 )
 from warnings import warn
 
-from qiskit_ibm_provider import IBMProvider  # type: ignore
-from qiskit_ibm_provider.exceptions import IBMProviderError  # type: ignore
-from qiskit.primitives import SamplerResult  # type: ignore
+import numpy as np
+
+from qiskit.primitives import PrimitiveResult, SamplerResult  # type: ignore
 
 
 # RuntimeJob has no queue_position attribute, which is referenced
@@ -45,12 +45,12 @@ from qiskit.result.distributions import QuasiDistribution  # type: ignore
 from qiskit_ibm_runtime import (  # type: ignore
     QiskitRuntimeService,
     Session,
-    Options,
-    Sampler,
+    SamplerOptions,
+    SamplerV2,
     RuntimeJob,
 )
 
-from pytket.circuit import Circuit, OpType
+from pytket.circuit import Bit, Circuit, OpType
 from pytket.backends import Backend, CircuitNotRunError, CircuitStatus, ResultHandle
 from pytket.backends.backendinfo import BackendInfo
 from pytket.backends.backendresult import BackendResult
@@ -96,7 +96,7 @@ from .ibm_utils import _STATUS_MAP, _batch_circuits
 from .config import QiskitConfig
 
 if TYPE_CHECKING:
-    from qiskit_ibm_provider.ibm_backend import IBMBackend as _QiskIBMBackend  # type: ignore
+    from qiskit.providers.backend import BackendV1  # type: ignore
 
 _DEBUG_HANDLE_PREFIX = "_MACHINE_DEBUG_"
 
@@ -125,21 +125,10 @@ def _save_ibmq_auth(qiskit_config: Optional[QiskitConfig]) -> None:
     token = None
     if qiskit_config is not None:
         token = qiskit_config.ibmq_api_token
-    try:
-        if token is not None:
-            IBMProvider.save_account(token, overwrite=True)
-            IBMProvider()
-        else:
-            IBMProvider()
-    except:
-        if token is not None:
-            IBMProvider.save_account(token, overwrite=True)
-            IBMProvider()
-        else:
-            raise NoIBMQCredentialsError()
-    if not QiskitRuntimeService.saved_accounts():
-        if token is not None:
-            QiskitRuntimeService.save_account(channel="ibm_quantum", token=token)
+    if token is not None and not QiskitRuntimeService.saved_accounts():
+        QiskitRuntimeService.save_account(
+            channel="ibm_quantum", token=token, overwrite=True
+        )
 
 
 def _get_primitive_gates(gateset: Set[OpType]) -> Set[OpType]:
@@ -149,6 +138,12 @@ def _get_primitive_gates(gateset: Set[OpType]) -> Set[OpType]:
         return {OpType.X, OpType.SX, OpType.Rz, OpType.ECR}
     else:
         return gateset
+
+
+def _int_from_readout(readout: np.ndarray) -> int:
+    # Weird mixture of big- and little-endian here.
+    n_bytes = len(readout)
+    return sum(int(x) << (8 * (n_bytes - 1 - i)) for i, x in enumerate(readout))
 
 
 class IBMQBackend(Backend):
@@ -166,10 +161,15 @@ class IBMQBackend(Backend):
     :param monitor: Use the IBM job monitor. Defaults to True.
     :type monitor: bool, optional
     :raises ValueError: If no IBMQ account is loaded and none exists on the disk.
-    :param provider: An IBMProvider
-    :type provider: Optional[IBMProvider]
+    :param service: A QiskitRuntimeService
+    :type service: Optional[QiskitRuntimeService]
     :param token: Authentication token to use the `QiskitRuntimeService`.
     :type token: Optional[str]
+    :param sampler_options: A customised `qiskit_ibm_runtime` `SamplerOptions` instance.
+        See the Qiskit documentation at
+        https://docs.quantum.ibm.com/api/qiskit-ibm-runtime/qiskit_ibm_runtime.options.SamplerOptions
+        for details and default values.
+    :type sampler_options: Optional[SamplerOptions]
     """
 
     _supports_shots = False
@@ -182,17 +182,18 @@ class IBMQBackend(Backend):
         backend_name: str,
         instance: Optional[str] = None,
         monitor: bool = True,
-        provider: Optional["IBMProvider"] = None,
+        service: Optional[QiskitRuntimeService] = None,
         token: Optional[str] = None,
+        sampler_options: SamplerOptions = None,
     ):
         super().__init__()
         self._pytket_config = QiskitConfig.from_default_config_file()
-        self._provider = (
-            self._get_provider(instance=instance, qiskit_config=self._pytket_config)
-            if provider is None
-            else provider
+        self._service = (
+            self._get_service(instance=instance, qiskit_config=self._pytket_config)
+            if service is None
+            else service
         )
-        self._backend: "_QiskIBMBackend" = self._provider.get_backend(backend_name)
+        self._backend: "BackendV1" = self._service.get_backend(backend_name)
         config: QasmBackendConfiguration = self._backend.configuration()
         self._max_per_job = getattr(config, "max_experiments", 1)
 
@@ -214,29 +215,22 @@ class IBMQBackend(Backend):
         # cache of results keyed by job id and circuit index
         self._ibm_res_cache: Dict[Tuple[str, int], Counter] = dict()
 
+        if sampler_options is None:
+            sampler_options = SamplerOptions()
+        self._sampler_options = sampler_options
+
         self._MACHINE_DEBUG = False
 
     @staticmethod
-    def _get_provider(
+    def _get_service(
         instance: Optional[str],
         qiskit_config: Optional[QiskitConfig],
-    ) -> "IBMProvider":
+    ) -> QiskitRuntimeService:
         _save_ibmq_auth(qiskit_config)
-        try:
-            if instance is not None:
-                provider = IBMProvider(instance=instance)
-            else:
-                provider = IBMProvider()
-        except IBMProviderError as err:
-            logging.warn(
-                (
-                    "Provider was not specified enough, specify hub,"
-                    "group and project correctly (check your IBMQ account)."
-                )
-            )
-            raise err
-
-        return provider
+        if instance is not None:
+            return QiskitRuntimeService(channel="ibm_quantum", instance=instance)
+        else:
+            return QiskitRuntimeService(channel="ibm_quantum")
 
     @property
     def backend_info(self) -> BackendInfo:
@@ -317,17 +311,16 @@ class IBMQBackend(Backend):
 
     @classmethod
     def available_devices(cls, **kwargs: Any) -> List[BackendInfo]:
-        provider: Optional["IBMProvider"] = kwargs.get("provider")
-        if provider is None:
-            if kwargs.get("instance") is not None:
-                provider = cls._get_provider(
-                    instance=kwargs.get("instance"), qiskit_config=None
-                )
+        service: Optional[QiskitRuntimeService] = kwargs.get("service")
+        if service is None:
+            instance = kwargs.get("instance")
+            if instance is not None:
+                service = cls._get_service(instance=instance, qiskit_config=None)
             else:
-                provider = IBMProvider()
+                service = QiskitRuntimeService(channel="ibm_quantum")
 
         backend_info_list = []
-        for backend in provider.backends():
+        for backend in service.backends():
             config = backend.configuration()
             props = backend.properties()
             backend_info_list.append(cls._get_backend_info(config, props))
@@ -503,7 +496,11 @@ class IBMQBackend(Backend):
                 apply the pytket ``SimplifyInitial`` pass to improve
                 fidelity of results assuming all qubits initialized to zero
                 (bool, default False)
-
+            * `sampler_options`:
+                A customised `qiskit_ibm_runtime` `SamplerOptions` instance. See
+                the Qiskit documentation at
+                https://docs.quantum.ibm.com/api/qiskit-ibm-runtime/qiskit_ibm_runtime.options.SamplerOptions
+                for details and default values.
         """
         circuits = list(circuits)
 
@@ -519,6 +516,10 @@ class IBMQBackend(Backend):
         postprocess = kwargs.get("postprocess", False)
         simplify_initial = kwargs.get("simplify_initial", False)
 
+        sampler_options: SamplerOptions = kwargs.get("sampler_options")
+        if sampler_options is None:
+            sampler_options = self._sampler_options
+
         batch_id = 0  # identify batches for debug purposes only
         for (n_shots, batch), indices in zip(circuit_batches, batch_order):
             for chunk in itertools.zip_longest(
@@ -532,11 +533,14 @@ class IBMQBackend(Backend):
 
                 qcs, ppcirc_strs = [], []
                 for tkc in batch_chunk:
+                    tkc1 = tkc.copy()
+                    # Flatten bits to default register in lexicographic order:
+                    tkc1.rename_units({bit: Bit(i) for i, bit in enumerate(tkc1.bits)})
                     if postprocess:
-                        c0, ppcirc = prepare_circuit(tkc, allow_classical=False)
+                        c0, ppcirc = prepare_circuit(tkc1, allow_classical=False)
                         ppcirc_rep = ppcirc.to_dict()
                     else:
-                        c0, ppcirc_rep = tkc, None
+                        c0, ppcirc_rep = tkc1, None
                     if simplify_initial:
                         SimplifyInitial(
                             allow_classical=False, create_all_qubits=True
@@ -552,15 +556,8 @@ class IBMQBackend(Backend):
                             ppcirc_strs[i],
                         )
                 else:
-                    options = Options()
-                    options.optimization_level = 0
-                    options.resilience_level = 0
-                    options.transpilation.skip_transpilation = True
-                    options.execution.shots = n_shots
-                    sampler = Sampler(session=self._session, options=options)
-                    job = sampler.run(
-                        circuits=qcs,
-                    )
+                    sampler = SamplerV2(session=self._session, options=sampler_options)
+                    job = sampler.run(qcs, shots=n_shots)
                     job_id = job.job_id()
                     for i, ind in enumerate(indices_chunk):
                         handle_list[ind] = ResultHandle(
@@ -628,10 +625,19 @@ class IBMQBackend(Backend):
                         sleep(10)
 
                 res = job.result(timeout=kwargs.get("timeout", None))
-            for circ_index, (r, d) in enumerate(zip(res.quasi_dists, res.metadata)):
-                self._ibm_res_cache[(jobid, circ_index)] = Counter(
-                    {n: int(0.5 + d["shots"] * p) for n, p in r.items()}
-                )
+            if isinstance(res, SamplerResult):
+                # TODO Is this code still reachable?
+                for circ_index, (r, d) in enumerate(zip(res.quasi_dists, res.metadata)):
+                    self._ibm_res_cache[(jobid, circ_index)] = Counter(
+                        {n: int(0.5 + d["shots"] * p) for n, p in r.items()}
+                    )
+            else:
+                assert isinstance(res, PrimitiveResult)
+                for circ_index, pub_result in enumerate(res._pub_results):
+                    readouts = pub_result.data.c.array
+                    self._ibm_res_cache[(jobid, circ_index)] = Counter(
+                        _int_from_readout(readout) for readout in readouts
+                    )
 
         counts = self._ibm_res_cache[cache_key]  # Counter[int]
         # Convert to `OutcomeArray`: